requirements.txt
click==8.1.7
commonroad_drivability_checker==2023.1
commonroad_reach==2023.1
geopandas==0.10.2
lgsvl==2021.1+30.g5aa0c06
lxml==4.9.3
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
omegaconf==2.3.0
ordered_set==4.1.0
osmium==3.7.0
Pillow==9.5.0
pyproj==3.2.1
PyQt5==5.15.10
PyYAML==6.0.1
scipy==1.7.3
Shapely==2.0.2
timeout_decorator==0.5.0
utm==0.7.0
